requirements.txt
streamlit
numpy
pandas
requests
datetime 
matplotlib.pyplot
matplotlib.dates